Refactor admin components to use shared patterns and fix code review issues
