+++ app/api/admin/logout/route.ts
@@ -3,7 +3,7 @@ import { cookies } from "next/headers"
 
 export async function POST() {
   try {
-    const cookieStore = cookies()
+    const cookieStore = await cookies()
     cookieStore.delete("admin-token")
 
     return NextResponse.json({
